Extract async step input snapshot population
